+++ src/test/java/OrderCreatingTest.java
@@ -1,6 +1,6 @@
-import io.restassured.RestAssured;
 import io.restassured.response.Response;
-import org.junit.Before;
+import model.Order;
+import org.apache.http.HttpStatus;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.junit.runners.Parameterized;
@@ -12,13 +12,12 @@ import static io.restassured.RestAssured.given;
 import static org.hamcrest.CoreMatchers.notNullValue;
 
 @RunWith(Parameterized.class)
-public class OrderCreatingTest {
+public class OrderCreatingTest extends BaseTest{
 
     private final List<String> colors;
 
     public OrderCreatingTest(List<String> colors) {
         this.colors = colors;
-
     }
 
     @Parameters
@@ -31,15 +30,17 @@ public class OrderCreatingTest {
         };
     }
 
-    @Before
-    public void setUp() {
-        RestAssured.baseURI = "http://qa-scooter.praktikum-services.ru/";
-    }
-
-
     @Test
     public void createOrder() {
-        Order order = new Order(colors);
+        Order order = new Order("Wolf",
+                "The Grey",
+                "Miracle forest",
+                "Under the Oak",
+                "+7 800 755 25 65",
+                6,
+                "2022-06-06",
+                "some comments",
+                colors);
 
         Response response = given()
                 .header("Content-type", "application/json")
@@ -50,98 +51,7 @@ public class OrderCreatingTest {
 
         response.then().assertThat().body("track", notNullValue())
                 .and()
-                .statusCode(201);
+                .statusCode(HttpStatus.SC_CREATED);
     }
 
-    private static class Order {
-
-        private String firstName = "Wolf";
-        private String lastName = "The Grey";
-        private String address = "Miracle forest";
-        private String metroStation = "Under the Oak";
-        private String phone = "+7 800 755 25 65";
-        private Integer rentTime = 6;
-        private String deliveryDate = "2022-06-06";
-        private String comment = "some comment";
-        private List<String> colors;
-
-        public Order() {
-        }
-
-        private Order(List<String> colors) {
-            this.colors = colors;
-        }
-
-        public List<String> getColors() {
-            return colors;
-        }
-
-        public void setColors(List<String> colors) {
-            this.colors = colors;
-        }
-
-        public String getFirstName() {
-            return firstName;
-        }
-
-        public void setFirstName(String firstName) {
-            this.firstName = firstName;
-        }
-
-        public String getLastName() {
-            return lastName;
-        }
-
-        public void setLastName(String lastName) {
-            this.lastName = lastName;
-        }
-
-        public String getAddress() {
-            return address;
-        }
-
-        public void setAddress(String address) {
-            this.address = address;
-        }
-
-        public String getMetroStation() {
-            return metroStation;
-        }
-
-        public void setMetroStation(String metroStation) {
-            this.metroStation = metroStation;
-        }
-
-        public String getPhone() {
-            return phone;
-        }
-
-        public void setPhone(String phone) {
-            this.phone = phone;
-        }
-
-        public Integer getRentTime() {
-            return rentTime;
-        }
-
-        public void setRentTime(Integer rentTime) {
-            this.rentTime = rentTime;
-        }
-
-        public String getDeliveryDate() {
-            return deliveryDate;
-        }
-
-        public void setDeliveryDate(String deliveryDate) {
-            this.deliveryDate = deliveryDate;
-        }
-
-        public String getComment() {
-            return comment;
-        }
-
-        public void setComment(String comment) {
-            this.comment = comment;
-        }
-    }
 }
