+++ src/test/java/OrdersListTest.java
@@ -1,19 +1,12 @@
-import io.restassured.RestAssured;
-import org.junit.Before;
+import org.apache.http.HttpStatus;
 import org.junit.Test;
 
 import java.util.List;
 
 import static io.restassured.RestAssured.given;
-import static org.hamcrest.CoreMatchers.equalTo;
 import static org.hamcrest.Matchers.isA;
 
-public class OrdersListTest {
-
-    @Before
-    public void setUp() {
-        RestAssured.baseURI = "http://qa-scooter.praktikum-services.ru/";
-    }
+public class OrdersListTest extends BaseTest{
 
     @Test
     public void getListOfOrder() {
@@ -21,6 +14,6 @@ public class OrdersListTest {
                 .get("/api/v1/orders")
                 .then().assertThat().body("orders", isA(List.class))
                 .and()
-                .statusCode(200);
+                .statusCode(HttpStatus.SC_OK);
     }
 }
