+++ src/test/java/model/Courier.java
@@ -1,4 +1,4 @@
-package json;
+package model;
 
 public class Courier {
